InventoryAppExample/requirements.txt
appdirs==1.4.4
avro==1.10.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.0
distlib==0.3.1
docopt==0.6.2
filelock==3.0.12
Flask==1.1.2
grpc-tools==0.0.1
grpcio==1.39.0
grpcio-tools==1.39.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
pipreqs==0.4.10
protobuf==3.17.3
requests==2.25.1
six==1.15.0
urllib3==1.26.6
virtualenv==20.4.4
Werkzeug==2.0.0
yarg==0.1.9
